Loading char data in background thread
When  joined_game is called there is no guarantie that we will have a valid chardata.

to adress this, plugin now tries to load char data multiple times in the background thread.

this ensures that correct config data is always loaded when the game starts for the plug in WHILE ensuring the conditions for other tasks are not triggered.

plugin will only proceed with the other tasks such as preperation etc, only after the loading task is complete.